refactor: Update form settings persistance
