+++ packages/form-builder/src/settings/donation-goal/index.js
@@ -1,4 +1,4 @@
-import {useFormSettings} from "../../context/formSettings";
+import {useFormSettings} from "../context";
 import {__} from "@wordpress/i18n";
 import {
     __experimentalNumberControl as NumberControl,
