+++ packages/form-builder/src/settings/form-fields/index.js
@@ -1,4 +1,4 @@
-import {useFormSettings} from "../../context/formSettings";
+import {useFormSettings} from "../context";
 import {__} from "@wordpress/i18n";
 import {PanelBody, PanelRow, SelectControl, ToggleControl} from "@wordpress/components";
